refactor: split _shell.py (1,216 lines) into focused modules (fixes #833)
Split the monolithic ShellMixin into a _shell/ package with 6 sub-mixins:
- _app.py: list_apps, launch_app, quit_app
- _menu.py: menu enumeration via Win32 API and UIA fallback
- _dialog.py: open_uri, dialog detection and interaction
- _taskbar.py: taskbar listing and clicking
- _tray.py: system tray icon enumeration and interaction
- _desktop.py: virtual desktop management via pyvda

ShellMixin in __init__.py composes all sub-mixins via MRO.
No public API changes — all 30 methods remain on ShellMixin.
